clean: ParamsとEventTypeをmSlackに切り分ける #6
