+++ lib/m_slack/event_types.go
@@ -1,0 +1,9 @@
+package mSlack
+
+type EventType string
+
+const (
+	AppMentionEvent EventType = "AppMentionEvent"
+	SlashEvent      EventType = "SlashEvent"
+	URLVerification EventType = "URLVerification"
+)
