+++ lib/m_slack/params.go
@@ -1,0 +1,10 @@
+package mSlack
+
+type EventParams struct {
+	Token       string
+	RequestBody []byte
+	Type        EventType
+	RequestKey  string
+	UserID      string
+	ChannelID   string
+}
